rename mosque data service
